phase4_5_rag_coaching/requirements.txt
sentence-transformers==2.7.0
faiss-cpu
groq==0.11.0
python-dotenv==1.0.1
numpy==1.26.4
